// src/bot/types/health-event.interface.ts
export interface HealthEvent {
  nombre_del_evento: string;
  urbano: number;
  rural: number;
  primera_infancia: number;
  infancia: number;
  adolescencia: number;
  juventud: number;
  adulto_j_ven: number;
  adulto_mayor: number;
  femenino: number;
  masculino: number;
  total_de_eventos: number;
}